Updated plotter script to respect column names.
Note: Currently, it assumes that colheaders of NWN voltage and NWN current are consistent.
If you change colheader names, it will no longer work.
I may update this in the future, but it is rather unlikely at the moment. Simply rememer to name them NW1 current/NW1 voltage and so on. Swapping these around no longer is a problem though.