Many changes, hopefully  nothing feature breaking
started work on a SEM stitching function. most of those implementations are in a document not in the repo, but it will be adapted later. A function to easily remove annotations was added (without rescale/filtering) and some changes to the base code of the other functions were implemented too.